requirements.txt
gradio
python-dotenv
llama-index
llama-parse
llama-index-llms-groq
llama-index-embeddings-cohere
torch
transformers